+++ static_precompiler/tests/test_management.py
@@ -37,6 +37,7 @@ def test_compilestatic_command(monkeypatch, tmpdir):
         "scripts/test.js",
         "styles/less/imported.css",
         "styles/less/test.css",
+        "styles/sass/precision.css",
         "styles/sass/test.css",
         "styles/stylus/A.css",
         "styles/stylus/B/C.css",